requirements.txt
aniso8601==9.0.1
attrs==23.2.0
bcrypt==4.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dlib==19.24.2
docutils==0.20.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lmdb==1.4.1
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
methodtools==0.4.7
nltk==3.8.1
numpy==1.24.4
opencv-python==4.10.0.84
pandas==2.0.3
path==16.10.0
path.py==12.5.0
pdf2image==1.9.0
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
pip==24.2
pkgutil_resolve_name==1.3.10
preview_generator==0.29
previewr==0.4.0
py-lz4framed==0.14.0
py-rouge==1.1
pycparser==2.22
pyexifinfo==0.4.0
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyPDF2==2.11.1
pypdfium2==4.30.0
pyrouge==0.1.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-magic==0.4.27
pytorch-pretrained-bert==0.4.0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
scikit-learn==1.3.2
scipy==1.10.1
setuptools==71.0.4
six==1.16.0
threadpoolctl==3.5.0
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Wand==0.6.13
Werkzeug==3.0.3
wheel==0.43.0
wirerope==0.4.7
zipp==3.19.2
zstandard==0.23.0

face_recognition~=1.3.0